requirements.txt
requests
discord.py
datetime
python-dotenv